Add a check for keystone expired tokens buildup.
More than 1K is too many when cleaning every 5 minutes, so warn about
it. The admin may have to go in and more aggressively clean or figure
out why the cleaning isn't completing as expected.

This uses a lot of keystone-manage code to interact with the database,
and monkeypatches itself into that code, hence why it needs to be
executed from the keystone venv.

It also only works with sql backend of tokens, but that's all we
support.