Name checks based on example number rather than line number
When looping over lists during similar ServerSpec checks, if you do not
register each check rather than the line number, then only the last item
in the list matters in the sensu dashboard because it will have the same
check name as the other list items. This cause those results to be overwritten.

This commit changes the check naming convention from <filename>_<linenumber> to
<filename>_<checknumber>.