check-serverspec.rb: only output summary and failing items.
Do not output a sensu check result for every serverspec check with a
unique hash. This causes a massive explosion of checks to be processed
and transmitted and an even larger amount of state which must be stored.

Instead, output a single summary with a list of all failing items, the
file they are defined in at which line number and why the check failed.